+++ MCSC/main.swift
@@ -1,5 +1,7 @@
 import Cocoa
 
-let delegate = AppDelegate()
-NSApplication.shared.delegate = delegate
+MainActor.assumeIsolated {
+    let delegate = AppDelegate()
+    NSApplication.shared.delegate = delegate
+}
 _ = NSApplicationMain(CommandLine.argc, CommandLine.unsafeArgv)